better legends and merge pit scouting into data dash
