+++ src/components/App/Loader/Loader.jsx
@@ -41,9 +41,7 @@ export function Loader(props) {
 			// else if testing
 			} else {
 				// load events
-				readEvents().then(events => {
-					console.log('read events: ', events);
-				}, e => {
+				readEvents().then(() => {}, e => {
 					console.error('error loading events from local db: ', e);
 				});
 			}
