+++ src/components/pages/Data/Data-content.jsx
@@ -5,6 +5,7 @@ import { Typography } from '@material-ui/core';
 
 import { Selector } from './Selector.jsx';
 import { DataCard } from './DataCard.jsx';
+import { PitCard } from './PitCard.jsx';
 
 const useStyles = makeStyles(theme => ({
 	root: {
@@ -15,12 +16,15 @@ const useStyles = makeStyles(theme => ({
 		display: 'flex',
 		flexWrap: 'wrap',
 		justifyContent: 'center',
+		alignItems: 'start',
+	},
+	dataCard: {
 	},
 }));
 
 export function DataContent(props) {
 	const classes = useStyles();
-	const { template, events, processedTeams, rawRuns } = props;
+	const { template, events, processedTeams, rawRuns, updatePit, pits } = props;
 
 	let startedEvents = events.filter(event => event.startDate < Date.now()).sort((a, b) => a.startDate - b.startDate);
 	
@@ -33,7 +37,6 @@ export function DataContent(props) {
 		initialSelection.event = startedEvents.length > 0 ? startedEvents[startedEvents.length - 1].key : events.sort((a, b) => a.startDate - b.startDate)[0].key;
 		
 		let availableTeams = events.filter(event => event.key === initialSelection.event)[0].teams ? events.filter(event => event.key === initialSelection.event)[0].teams.sort((a, b) => a - b) : [];
-		console.log(availableTeams)
 		initialSelection.team = availableTeams.length > 0 ? availableTeams[0] : null;
 	}
 	
@@ -42,6 +45,7 @@ export function DataContent(props) {
 	const teams = events.filter(event => event.key === selection.event)[0].teams ? events.filter(event => event.key === selection.event)[0].teams : [];
 	const selectedObj = processedTeams.filter(obj => (obj.team === selection.team && obj.event === selection.event))[0];
 	const selectedRuns = rawRuns.filter(obj => (obj.team === selection.team && obj.event === selection.event));
+	const selectedPit = pits.filter(pit => (pit.team === selection.team && pit.event === selection.event))[0];
 
 	return (
 		<div className={classes.root}>
@@ -51,13 +55,14 @@ export function DataContent(props) {
 				{teams.length < 1 &&
 					<Typography variant='h6'>No teams available!</Typography>
 				}
-				{(teams.length > 0 && (!selectedObj || (selectedObj && !selectedObj.data))) &&
-					<Typography variant='h6'>No data available!</Typography>
+				
+				{ teams.length > 1 &&
+					<PitCard template={template} event={selection.event} team={selection.team} submit={updatePit} readPit={selectedPit} />
 				}
 
 				{/* if an object matches the selection AND that object has a dataset */}
 				{selectedObj && selectedObj.data && Object.keys(selectedObj.data).map(key => (
-					<DataCard key={key} template={template} topKey={key} processedObject={selectedObj} teamsRuns={selectedRuns} />
+					<DataCard key={key} className={classes.dataCard} template={template} topKey={key} processedObject={selectedObj} teamsRuns={selectedRuns} />
 				))}
 			</div>
 		</div>
