--- src/components/pages/Pit/index.js
@@ -1,2 +1,0 @@
-import Pit from './Pit-container';
-export default Pit;
