+++ src/utils/database.js
@@ -70,6 +70,31 @@ export function readRuns(event, team) {
 	});
 };
 
+export function readTeams(event, team) {
+	return new Promise((resolve, reject) => {
+		const db = new Dexie('strangescout');
+		versions.forEach(version => {
+			db.version(version.version).stores(version.stores);
+		});
+
+		db.teams.toArray().then(docs => {
+			let finalDocs = docs;
+			
+			if (team) {
+				finalDocs = finalDocs.filter(doc => doc.team === team);
+			}
+			if (event) {
+				finalDocs = finalDocs.filter(doc => doc.event === event);
+			}
+
+			resolve(finalDocs);
+		}, e => {
+			console.error('error reading pits from local db: ', e);
+			reject(e);
+		});
+	});
+};
+
 export function readProcessedTeams() {
 	return new Promise((resolve, reject) => {
 		const db = new Dexie('strangescout');
